implementations/moe-model/requirements.txt
torch
pandas
scikit-learn
networkx
spacy
flask
jupyter
pdfplumber
sentence-transformers
fpdf
numpy
pypdf2
faiss-cpu